Changed: Scan() return "none" if not holding. Changed: SpeedCommand now accepts a float value (e.g. speed 0.5f) Changed: changed FilterDataIntoMessage() to a simpler function, which was needed for the (python) libraries.

+++ Assets/Models/CommandBuilder.cs
@@ -33,7 +33,7 @@ namespace Assets.Models
             }
             else if (instruction == "speed")
             {
-                return new SpeedCommand(int.Parse(parameter));
+                return new SpeedCommand((int)float.Parse(parameter));
             }
             else if (instruction == "scan")
             {
